v5.x-refactor-tests: removed RuntimeCosting.scala
