+++ sigmastate/src/main/scala/sigmastate/eval/TreeBuilding.scala
@@ -10,11 +10,12 @@ import sigmastate.serialization.ConstantStore
 
 import scala.collection.mutable.ArrayBuffer
 import SType._
+import scalan.SigmaLibrary
 import sigmastate.basics.DLogProtocol.ProveDlog
 import sigmastate.basics.ProveDHTuple
 import sigmastate.lang.SigmaTyper
 
-trait TreeBuilding extends RuntimeCosting { IR: IRContext =>
+trait TreeBuilding extends SigmaLibrary { IR: IRContext =>
   import Liftables._
   import Context._
   import SigmaProp._
