+++ sigmastate/src/test/scala/sigmastate/utxo/examples/CoinEmissionSpecification.scala
@@ -28,8 +28,6 @@ class CoinEmissionSpecification extends SigmaTestingCommons
     // uncomment if you want to log script evaluation
     // override val okPrintEvaluatedEntries = true
     saveGraphsInFile = false
-    outputEstimatedCost = false
-    outputComputedResults = false
   }
 
   private val reg1 = ErgoBox.nonMandatoryRegisters.head
